+++ packages/user-interface/src/pages/account/account-page.mock.tsx
@@ -1,18 +1,14 @@
 import React, {ReactElement} from 'react';
 import {AccountPage} from './account-page';
-import {TestProviderWrapper} from '../../testUtils/TestProviderWrapper/TestProviderWrapper';
+import {TestProviderWrapper} from '../../testUtils/TestProviderWrapper';
 import {mocksRequestBurgerGegevens} from './account-page-requests.mock';
 
 const route = '/zaken/zaak?id=82cb13cf-d2f9-4e3e-ac07-751373035ecb';
 
-const AccountPageTestComponent = ({}): ReactElement => {
-  return <AccountPage showNotificationSubSection="false" />;
-};
-
 export const MockAccountPage = (): ReactElement => {
   return (
     <TestProviderWrapper mocks={mocksRequestBurgerGegevens} route={route}>
-      <AccountPageTestComponent />
+      <AccountPage showNotificationSubSection="false" />
     </TestProviderWrapper>
   );
 };
