+++ packages/user-interface/src/pages/account/account-page.test.tsx
@@ -27,7 +27,7 @@ describe('AccountPage', () => {
     expect(screen.getByText('Kooyman')).toBeVisible();
     expect(screen.getByText('vrouw')).toBeVisible();
     expect(screen.getByText('999991954')).toBeVisible();
-    expect(screen.getByText('3 maart 2003')).toBeVisible();
+    expect(screen.getByText('3 March 2003')).toBeVisible();
     expect(screen.getByText('Nederland')).toBeVisible();
     expect(screen.getByText('Leyweg 61e')).toBeVisible();
     expect(screen.getByText("2545CC 's-Gravenhage")).toBeVisible();
