+++ packages/user-interface/src/pages/case/case-page.test.tsx
@@ -1,4 +1,4 @@
-import {render, screen, waitForElementToBeRemoved} from '@testing-library/react';
+import {Matcher, render, screen, waitForElementToBeRemoved} from '@testing-library/react';
 import {
   MockCasePage,
   MockCasePageWithoutContactMoments,
@@ -6,6 +6,8 @@ import {
 } from './case-page.mock';
 
 describe('CasePage', () => {
+  let pdfFile: Matcher;
+
   beforeAll(() => {
     Object.defineProperty(window, 'matchMedia', {
       writable: true,
@@ -20,32 +22,36 @@ describe('CasePage', () => {
         dispatchEvent: jest.fn(),
       })),
     });
+    pdfFile = 'Certificaat WWJB.pdf (pdf, 60.2 kB)';
   });
 
   it('should render with all elements present', async () => {
     render(MockCasePage());
-    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Aan het laden'));
-    expect(screen.queryByText('Certificaat WWJB.pdf (pdf, 60,2 kB)')).toBeVisible();
-    expect(screen.getByText('In behandeling genomen')).toBeVisible();
+    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Loading'));
+
+    expect(screen.getByText(pdfFile)).toBeVisible();
+    expect(screen.getByText('Notice of objection')).toBeVisible();
+
     expect(screen.getByText('ZAAK-2023-0000007947')).toBeVisible();
     expect(screen.getByRole('table')).toBeVisible();
     expect(screen.getByText('Dit is een sms')).toBeVisible();
-    expect(screen.queryByText('Eerdere contactmomenten')).toBeVisible();
+    expect(screen.queryByText('Previous contact moments')).toBeVisible();
   });
 
   it('should render without any documents present', async () => {
     render(MockCasePageWithoutDocuments());
-    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Aan het laden'));
-    expect(screen.queryByText('Certificaat WWJB.pdf (pdf, 60,2 kB)')).toBeNull();
-    expect(screen.queryByText('Er zijn geen documenten.')).toBeVisible();
-    expect(screen.queryByText('Eerdere contactmomenten')).toBeVisible();
+    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Loading'));
+    expect(screen.queryByText(pdfFile)).toBeNull();
+    expect(screen.queryByText('There are no documents.')).toBeVisible();
+    expect(screen.queryByText('Previous contact moments')).toBeVisible();
   });
 
   it('should render without any contactmoments present', async () => {
     render(MockCasePageWithoutContactMoments());
-    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Aan het laden'));
-    expect(screen.getByText('Documenten'));
-    expect(screen.queryByText('Certificaat WWJB.pdf (pdf, 60,2 kB)')).toBeVisible();
-    expect(screen.queryByText('Eerdere contactmomenten')).toBeNull();
+    await waitForElementToBeRemoved(() => screen.getAllByLabelText('Loading'));
+    expect(screen.getByText('Documents'));
+
+    expect(screen.queryByText(pdfFile)).toBeVisible();
+    expect(screen.queryByText('Previous contact moments')).toBeNull();
   });
 });
