+++ packages/user-interface/src/pages/edit-account/edit-account-page.test.tsx
@@ -2,10 +2,10 @@ import {fireEvent, render, screen} from '@testing-library/react';
 import {MockEditAccountPage} from './edit-account-page.mock';
 
 describe('EditAccountPage', () => {
-  const errorText = 'een geldig telefoonnummer bestaat uit 10 cijfers';
+  const errorText = 'a valid phone number consists of 10 digits';
   const inputField = () => screen.getByRole('textbox');
   const errorTextP = () => screen.queryByText(errorText);
-  const saveButton = () => screen.getByText('Opslaan');
+  const saveButton = () => screen.getByText('Save');
 
   beforeEach(() => {
     render(MockEditAccountPage());
