--- packages/user-interface/src/testUtils/custom-messages/nl-nl.ts
@@ -1,32 +1,0 @@
-import {DEFAULT_LOCALES, Messages} from '@nl-portal/nl-portal-localization';
-
-export const NL_NL_MESSAGES: Messages = {
-  [DEFAULT_LOCALES.DUTCH]: {
-    'app.appName': 'MijnDenHaag',
-    'case.bezwaar-behandelen.title': 'Bezwaarschrift',
-    'case.bezwaar-behandelen.status.intake-afgerond': 'Intake afgerond',
-    'case.bezwaar-behandelen.status.indieningsvereisten-getoetst': 'Indieningsvereisten getoetst',
-    'case.bezwaar-behandelen.status.bezwaar-beoordeeld': 'Bezwaar beoordeeld',
-    'case.bezwaar-behandelen.status.hoorzitting-gehouden': 'Hoorzitting gehouden',
-    'case.bezwaar-behandelen.status.concept-besluit-opgesteld': 'Concept besluit opgesteld',
-    'case.bezwaar-behandelen.status.besluit-vastgesteld': 'Besluit vastgesteld',
-    'case.bezwaar-behandelen.status.zaak-afgerond': 'Zaak afgerond',
-    'case.B0756.title': 'Bezwaarschrift',
-    'case.B0756.status.voorstel-voor-besluitvorming-opgesteld':
-      'Voorstel voor besluitvorming opgesteld',
-    'case.B0756.status.besluitvorming-afgerond': 'Besluitvorming afgerond',
-    'case.B0756.status.onderzoek-afgerond': 'Onderzoek afgerond',
-    'case.B0756.status.afgehandeld': 'Afgehandeld',
-    'case.B0756.status.in-behandeling-genomen': 'In behandeling genomen',
-    'case.B0756.status.geaccepteerd': 'Geaccepteerd',
-    'case.B0756.status.geregistreerd': 'Geregistreerd',
-    'case.test-id.title': 'Testzaak',
-    'case.test-id.status.aangemaakt': 'Aangemaakt',
-    'footerColumns.theHague': 'Den Haag',
-    'footerColumns.disclaimers': 'Verklaringen',
-    'footerLinks.goToTheHague': 'Ga naar denhaag.nl',
-    'footerLinks.accessibility': 'Toegankelijkheidsverklaring',
-    'footerLinks.dataProtection': 'Verklaring inzake Gegevensbescherming',
-    'footerLinks.proclaimer': 'Proclaimer',
-  },
-};
